Refactoring 2.0
