+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "megapicker",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "A CLI to help you create different projects without looking up the doc command to find the installation commands",
   "main": "index.js",
   "type": "module",
